+++ packages/xchain-bitcoin/src/sochain-api.ts
@@ -1,4 +1,4 @@
-import { Network } from '@xchainjs/xchain-client'
+import { Network, TxHash } from '@xchainjs/xchain-client'
 import { BaseAmount, assetAmount, assetToBase } from '@xchainjs/xchain-util'
 import axios from 'axios'
 
@@ -150,6 +150,40 @@ export const getIsTxConfirmed = async ({ sochainUrl, network, hash }: TxHashPara
 }
 
 /**
+ * List of confirmed txs
+ *
+ * Stores a list of confirmed txs (hashes) in memory to avoid requesting same data
+ */
+const confirmedTxs: Array<TxHash> = []
+
+/**
+ * Helper to get `confirmed` status of a tx.
+ *
+ * It will get it from cache or try to get it from Sochain (if not cached before)
+ */
+export const getConfirmedTxStatus = async ({
+  txHash,
+  sochainUrl,
+  network,
+}: {
+  sochainUrl: string
+  txHash: TxHash
+  network: Network
+}): Promise<boolean> => {
+  // try to get it from cache
+  if (confirmedTxs.includes(txHash)) return true
+  // or get status from Sochain
+  const { is_confirmed } = await await getIsTxConfirmed({
+    sochainUrl,
+    network,
+    hash: txHash,
+  })
+  // cache status
+  confirmedTxs.push(txHash)
+  return is_confirmed
+}
+
+/**
  * Get unspent txs and filter out pending UTXOs
  *
  * @see https://sochain.com/api#get-unspent-tx
@@ -174,13 +208,13 @@ export const getConfirmedUnspentTxs = async ({
 
   await Promise.all(
     txs.map(async (tx: BtcAddressUTXO) => {
-      const { is_confirmed: isTxConfirmed } = await getIsTxConfirmed({
+      const confirmed = await getConfirmedTxStatus({
         sochainUrl,
         network,
-        hash: tx.txid,
+        txHash: tx.txid,
       })
 
-      if (isTxConfirmed) {
+      if (confirmed) {
         confirmedUTXOs.push(tx)
       }
     }),
